Widen FontSize from int to float end-to-end (#4304)
TextRuntime.FontSize, BmfcSave.FontSize/GetFontCacheFileNameFor, SkiaGum/RaylibGum's Text.FontSize, and the shared inline-BBCode font-size stack now all take float, letting a fractional size (e.g. 18.5) render distinctly from its neighboring integers.

KernSmith rasterizes the fractional value natively (FontGeneratorOptions.Size is float on the FreeType backend Gum registers). The legacy bmfont.exe/GDI backend can't, so BmfcSave.Save() and the bmfont.exe size-estimation heuristic round to the nearest whole pixel via the new BmfcSave.GetGdiRoundedFontSize.

Tool property grid is untouched; this ships as a runtime-only API.

Closes #4304

+++ MonoGameGum.Tests/Runtimes/TextRuntimeTests.cs
@@ -275,8 +275,8 @@ char id=67 x=0 y=0 width={xadvance} height=13 xoffset=0 yoffset=4 xadvance={xadv
     {
         public BitmapFont? TryCreateFont(BmfcSave bmfcSave)
         {
-            int advance = bmfcSave.FontSize / 2;
-            BitmapFont font = new BitmapFont((Texture2D)null!, AbcFontData(advance, lineHeight: bmfcSave.FontSize));
+            int advance = (int)bmfcSave.FontSize / 2;
+            BitmapFont font = new BitmapFont((Texture2D)null!, AbcFontData(advance, lineHeight: (int)bmfcSave.FontSize));
             font.SetFontPattern(256, 256);
             return font;
         }
